requirements.txt
openai
os
requests
python-dotenv
